day_11/task_12/requirements.txt
crewai
streamlit
google-generativeai
python-dotenv
tavily-python
pymupdf
markdown